修改 连续截图 修复重启终端后台没有收到停止通讯命令

+++ app/src/main/java/com/wosplayer/command/operation/other/Command_CAPT.java
@@ -13,6 +13,8 @@ import android.view.ViewGroup;
 import com.wosplayer.Ui.element.definedView.Mtextscroll;
 import com.wosplayer.Ui.element.definedView.Mvideo;
 import com.wosplayer.app.AppTools;
+import com.wosplayer.app.BackRunner;
+import com.wosplayer.app.DisplayActivity;
 import com.wosplayer.app.Logs;
 import com.wosplayer.app.SystemConfig;
 import com.wosplayer.command.operation.interfaces.iCommand;
@@ -32,6 +34,7 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -41,51 +44,79 @@ import cn.trinea.android.common.util.ShellUtils;
  * Created by user on 2016/7/29.
  * catch screen
  */
-public class Command_CAPT implements iCommand {
+public class Command_CAPT  extends Thread implements iCommand {
     private static final String TAG = "截屏";
+    private TimeScreen mThread;
     @Override
-    public void execute(Activity activity, String param) {
+    public void execute(final Activity activity, String param) {
 //        String command = "screencap -p "+ basepath;
+        //参数 6 - 没过6秒截屏一次. 参数 0 停止截屏
+
         SystemConfig config = SystemConfig.get().read();
         String capturePath = config.GetStringDefualt("CapturePath","");
         String uploadUrl = config.GetStringDefualt("CaptureURL","");
         String terminalNo = config.GetStringDefualt("terminalNo","");
-        boolean isDelete = (config.GetIntDefualt("CaptureSave",0) == 0 ?true:false);
-        boolean isNotify = (config.GetIntDefualt("CaptureNoty",0) == 0 ?true:false);
-        liunxCommadScreen(activity,terminalNo,capturePath,uploadUrl,isDelete,isNotify);
+        boolean isDelete = (config.GetIntDefualt("CaptureSave",0) == 0) ;
+        boolean isNotify = (config.GetIntDefualt("CaptureNoty",0) == 0);
+        int mode = config.GetIntDefualt("CaptureMode",0);
+
+        if (param == null || param.equals("")){
+            liunxCommadScreen(activity,terminalNo,capturePath,uploadUrl,isDelete,isNotify,mode);
+            return;
+        }
+        int time = 0;
+        try {
+            time = Integer.parseInt(param);
+        } catch (NumberFormatException e) {
+        }
+        Logs.e(TAG,"定时截屏循环时间:"+time+" 秒/次");
+        //停止截图线程
+        if (mThread!=null){
+            mThread.stopThread();
+            mThread = null;
+        }
+        if (time > 0){
+            mThread = new TimeScreen(activity,capturePath,uploadUrl,terminalNo,isDelete,isNotify,mode,time);
+            mThread.startThread();
+            mThread.start();
+        }
     }
 
-    private synchronized void liunxCommadScreen(Activity activity,String terminalNo, String savePath, String url,boolean isdelete,boolean isnoty) {
-        Logs.d(TAG,"开始>> 保存:"+terminalNo+" - 本地截图:"+savePath+" - 上传地址: "+url);
-        String cmd = "screencap -p "+savePath;
-        ShellUtils.CommandResult result = ShellUtils.execCommand(cmd,true,true);
-        if (result.result == 0){
-            Logs.d(TAG,"liunx 命令(screencap -p) 截屏成功 - "+savePath);
-        }
-        catchScreen(activity,terminalNo,savePath,url);
+    private synchronized void liunxCommadScreen(Activity activity,String terminalNo, String savePath, String url,boolean isdelete,boolean isnoty,int mode) {
 
+        Logs.d(TAG,"开始>> 保存:"+terminalNo+" - 本地截图:"+savePath+" - 上传地址: "+url);
+        if (mode == 0){
+            String cmd = "screencap -p "+savePath;
+            ShellUtils.CommandResult result = ShellUtils.execCommand(cmd,true,true);
+            if (result.result == 0){
+                Logs.d(TAG,"liunx 命令(screencap -p) 截屏完成 - "+savePath);
+            }
+        }
+        catchScreen(activity,terminalNo,savePath,url,mode);
         //上传
         uploadImage(activity,terminalNo,savePath,url,isdelete,isnoty);
     }
 
-    private void catchScreen(Activity activity,String terminalNo,String savePath,String url){
+    private void catchScreen(Activity activity,String terminalNo,String savePath,String url,int mode){
+        if (mode == 0){
+            try {
+                File image = new File(savePath);
+                if (image.exists()){
+                    BitmapFactory.Options options = new BitmapFactory.Options();
+                    options.inJustDecodeBounds=true;
+                    FileInputStream fis = new FileInputStream(image);
+                    BitmapFactory.decodeStream(fis,null,options);
+                    if (options.outWidth*options.outHeight > 10) return;
+                }
 
-        try {
-            File image = new File(savePath);
-            if (image.exists()){
-                BitmapFactory.Options options = new BitmapFactory.Options();
-                options.inJustDecodeBounds=true;
-                FileInputStream fis = new FileInputStream(image);
-                BitmapFactory.decodeStream(fis,null,options);
-                if (options.outWidth*options.outHeight > 10) return;
+                if (cn.trinea.android.common.util.FileUtils.isFileExist(savePath)){
+                    cn.trinea.android.common.util.FileUtils.deleteFile(savePath);
+                }
+            } catch (FileNotFoundException e) {
+                Logs.e(TAG, "命令行截屏文件无效 : "+e.getMessage() );
             }
-
-            if (cn.trinea.android.common.util.FileUtils.isFileExist(savePath)){
-                cn.trinea.android.common.util.FileUtils.deleteFile(savePath);
-            }
-        } catch (FileNotFoundException e) {
-            Logs.e(TAG, "命令行截屏文件无效 : "+e.getMessage() );
         }
+
         Bitmap bgbmp = null;
             FileOutputStream fos = null;
         try {
@@ -243,10 +274,11 @@ public class Command_CAPT implements iCommand {
      */
     private void uploadImage(final Activity activity, String terminalNo, String filePath, String url, boolean isDelete,boolean isNotify){
         File image = new File(filePath);
+        if (!image.exists()){
+            Logs.e("截图文件["+filePath+"]未找到.");
+            return;
+        }
         try {
-            if (!image.exists()){
-               new IllegalStateException("截图文件["+filePath+"]未找到.");
-            }
             HttpClient httpclient = new DefaultHttpClient();
             HttpPost httppost = new HttpPost(url);
             MultipartEntity entity = new MultipartEntity();
@@ -262,13 +294,15 @@ public class Command_CAPT implements iCommand {
             if (!result.equals("1")) {
                 new IllegalStateException("服务器返回值"+result);
             }
-            if (!isNotify) return;
-            activity.runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    AppTools.Toals(activity,"上传截图完成");
-                }
-            });
+            if (isNotify) {
+                activity.runOnUiThread(new Runnable() {
+                    @Override
+                    public void run() {
+                        AppTools.Toals(activity,"上传截图完成");
+                    }
+                });
+            }
+
         } catch (Exception e) {
             Logs.e("上传截图失败:"+e.getMessage());
         }finally{
@@ -277,8 +311,52 @@ public class Command_CAPT implements iCommand {
     }
 
 
+    //循环截图线程
+    private class TimeScreen extends Thread{
+        private WeakReference<Activity> activity;
+        private String capturePath;
+        private String uploadUrl;
+        private String terminalNo;
+        private boolean isDelete ;
+        private boolean isNotify ;
+        private int mode ;
+        private int time;
+        private boolean flag;
 
+        public TimeScreen(Activity activity, String capturePath, String uploadUrl, String terminalNo, boolean isDelete, boolean isNotify, int mode, int time) {
+            this.activity = new WeakReference<Activity>(activity);
+            this.capturePath = capturePath;
+            this.uploadUrl = uploadUrl;
+            this.terminalNo = terminalNo;
+            this.isDelete = isDelete;
+            this.isNotify = isNotify;
+            this.mode = mode;
+            this.time = time;
+        }
 
+        public void startThread(){
+            flag = true;
+        }
+        public void stopThread(){
+            flag = false;
+        }
+        @Override
+        public void run() {
+            while(flag){
+                if (activity.get() == null){
+                    stopThread();
+                }else{
+                    liunxCommadScreen(activity.get(),terminalNo,capturePath,uploadUrl,isDelete,isNotify,mode);
+                    try {
+                        sleep(time * 1000);
+                    } catch (InterruptedException e) {
+                        stopThread();
+                    }
+                }
+
+            }
+        }
+    }
 
 
 }
